INI: make 'ini_*_serialize' API private
This allows to not expose "simplebuffer.h"
