+++ ini/ini_valueobj.c
@@ -26,6 +26,8 @@
 #include <ctype.h>
 #include "simplebuffer.h"
 #include "ref_array.h"
+#include "ini_configobj.h"
+#include "ini_config_priv.h"
 #include "ini_comment.h"
 #include "ini_defines.h"
 #include "ini_valueobj.h"
